% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Location.R
\name{get_all}
\alias{get_all}
\title{Lookup for IP address information}
\usage{
get_all(ip)
}
\arguments{
\item{ip}{IPv4 or IPv6 address}
}
\value{
Return all information about the IP address
}
\description{
Find the country, region, city, coordinates, zip code, ISP, domain name, timezone, connection speed, IDD code, area code, weather station code, weather station name, mobile, usage type, address type and IAB category that any IP address or host name originates from. The return values will be depending on the BIN data loaded.
}
\examples{
\dontrun{
get_all("8.8.8.8")
}

}
