% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2Location.R
\name{open}
\alias{open}
\title{Load IP2Location BIN data}
\usage{
open(bin_location)
}
\arguments{
\item{bin_location}{Absolute path of IP2Location BIN data}
}
\description{
Load the IP2Locaiton BIN data for lookup. Free IP2Location LITE data available for download at <https://lite.ip2location.com/>
}
\examples{
\donttest{
open("~/IP-COUNTRY.BIN")
}

}
